Add redirectUri to url regex for OAuth2 (#396)
* Add redirectUri to url regex for OAuth2

Keycloak, an OAuth2/OpenId Connect authentication server,
has multiple pages it redirects to during the login process
for the user. One such internal url it uses has `code=` as part
of the url. This causes the window to close prematurely and attempt
to get an access token. However, the code is wrong because it is
only an intermediaery step for Keycloak.

This fix will add the redirectUri to the regex. Insomnia will only attempt
to fetch the token AFTER keycloak has finished it's full login flow
and has redirected back to the redirectUri.

* Code review changes