Support toggling of auto beautify in editor. (#2305)
* Updated code editor to support toggling of beautify in editor.

Added a button to toggle beautify. When toggled text is 'beautified' immediately and when text is pasted into the editor or when the editor is blurred.

Removed _prettify method because it is no longer used.

* Changed toggle of auto beautify to dropdown selection.

The toggle has been replaced by a dropdown selection, the text in the beautify button will change accordingly. In manual mode, the button can still be clicked.

* Updated label of the beautify button.

* Changed caret up to cog icon + minor change to button label

* Changed cog icon to wrench icon.

Co-authored-by: Mike Ellan <52717970+sonicyeti@users.noreply.github.com>