Fix #4797 request switcher auto-focus from CodeMirror (#5002)
Fixes the input `.focus()` method call. The simplest fix was to revert the change in 3cf464d0593f4359c56e6443b8a6d19831b73127.

There do not seem to be any new problems in the current application from reverting this commit's relevant changes, indicating that the original cause of the bug is no longer present.